refactor(records): what "currently employed" means gets its own file
`relationship.go` crossed the 500-line cap the moment the currency helpers
landed in it, and the cap asks the right question: the two of them are one
concept with reach far outside that file. The write paths there decide the flag
with it, and four separate READS derive currency with it rather than trusting a
value written months earlier — which is exactly why it is worth finding on its
own rather than partway down the file that happens to have written it first.

No behaviour change; the block moved verbatim.
